+++ javaslang/src/main/java/javaslang/collection/SortedMap.java
@@ -5,7 +5,6 @@
  */
 package javaslang.collection;
 
-import javaslang.Function2;
 import javaslang.Tuple2;
 import javaslang.control.Option;
 
@@ -104,10 +103,10 @@ public interface SortedMap<K, V> extends Map<K, V> {
     <U> SortedMap<K, V> distinctBy(Function<? super Tuple2<K, V>, ? extends U> keyExtractor);
 
     @Override
-    SortedMap<K, V> drop(long n);
+    SortedMap<K, V> drop(int n);
 
     @Override
-    SortedMap<K, V> dropRight(long n);
+    SortedMap<K, V> dropRight(int n);
 
     @Override
     SortedMap<K, V> dropUntil(Predicate<? super Tuple2<K, V>> predicate);
@@ -143,7 +142,7 @@ public interface SortedMap<K, V> extends Map<K, V> {
     <C> Map<C, ? extends SortedMap<K, V>> groupBy(Function<? super Tuple2<K, V>, ? extends C> classifier);
 
     @Override
-    Iterator<? extends SortedMap<K, V>> grouped(long size);
+    Iterator<? extends SortedMap<K, V>> grouped(int size);
 
     @Override
     SortedMap<K, V> init();
@@ -208,10 +207,10 @@ public interface SortedMap<K, V> extends Map<K, V> {
     SortedMap<K, V> scan(Tuple2<K, V> zero, BiFunction<? super Tuple2<K, V>, ? super Tuple2<K, V>, ? extends Tuple2<K, V>> operation);
 
     @Override
-    Iterator<? extends SortedMap<K, V>> sliding(long size);
+    Iterator<? extends SortedMap<K, V>> sliding(int size);
 
     @Override
-    Iterator<? extends SortedMap<K, V>> sliding(long size, long step);
+    Iterator<? extends SortedMap<K, V>> sliding(int size, int step);
 
     @Override
     Tuple2<? extends SortedMap<K, V>, ? extends SortedMap<K, V>> span(Predicate<? super Tuple2<K, V>> predicate);
@@ -223,10 +222,10 @@ public interface SortedMap<K, V> extends Map<K, V> {
     Option<? extends SortedMap<K, V>> tailOption();
 
     @Override
-    SortedMap<K, V> take(long n);
+    SortedMap<K, V> take(int n);
 
     @Override
-    SortedMap<K, V> takeRight(long n);
+    SortedMap<K, V> takeRight(int n);
 
     @Override
     SortedMap<K, V> takeUntil(Predicate<? super Tuple2<K, V>> predicate);
